// src/app/model/consuntivo.ts
export class Consuntivo{
	data_consuntivo: 		Date;
	id_utente: 				any;
	nome_utente: 			string;
	id_cliente:				any;
	nome_cliente:			string;
	id_ambito: 				any;
	nome_ambito: 			string;
	id_macro_area: 			any;
	nome_macro_area: 		string;
	id_attivita: 			any;
	nome_attivita: 			string;
	id_tipo_deliverable: 	string;
	nome_tipo_deliverable: 	string;
	note: 					string;
	ore : 					number;
};